requirements.txt
phidata
google-generativeai
tavily-python
packaging
openai
pillow
streamlit
phi
pillow